Spell: Add check for single target to resist/miss channeled spell cancel Closes  https://github.com/cmangos/mangos-tbc/pull/645
I altered your PR a bit because:

Line with SPELL_MISS_IMMUNE2 should never be hit except for single target target. When spell is AOE, we do not supply unit target so check is redundant.
In second case, SPELL_MISS_RESIST - we want to just check if its the single target target. Like before, AOE does not supply that to the spell so its quicker.
The code is there for a spell like Mind Flay and nothing else really.